hive/services/rag-service/requirements.txt
fastapi>=0.104.1,<1.0.0
uvicorn>=0.24.0,<1.0.0
langchain>=0.0.335,<0.1.0
langchain-google-genai>=0.0.6
google-generativeai>=0.3.2
pgvector>=0.2.4
psycopg2-binary>=2.9.9
pypdf>=3.17.0
python-dotenv>=1.0.0
boto3>=1.29.0
pydantic>=2.5.0,<3.0.0
tiktoken>=0.5.1
tenacity>=8.2.3
sqlalchemy>=2.0.23,<3.0.0
